feat(pol): add support for matic/pol crypto fund PE-6722
